generator/requirements.txt
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
faker==25.9.1
